pyproject.toml
[tool.poetry]
name = "terragrunt-generator"
version = "0.1.0"
description = ""
authors = ["Chris <goabonga@pm.me>"]
readme = "README.md"
packages = [{include = "generator", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
python-hcl2 = "^3.0.5"
GitPython = "^3.1.27"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^22.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
terragrunt-generator = 'generator.main:main'
